+++ src/main/java/racegame/Constant.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racegame;
 
 public class Constant {
     public final static String INPUT_DELIMITER = ",";
